Add the session create action using log_in method
Find the user by specified email id. When the user is found
successfully, they are logged in and redirected to addresses
show page.

Else the alert message is displayed and login page is rendered again

Current user method is defined to find the user in the database
corresponding to the session id
The idea of defining current user is to retrieve the attributes of
logged in user to be used on subsequent pages