Allow users to edit their own information
Create correct_user and correct_address methods which will
check whether the current user? and current_address? is same as
logged in user

current_user? and current_address? are helper methods which
returns true if the user is current logged in user, otherwise
false and same with current_address

Otherwise if the user tries to edit some other user's information
, they will be redirected to welocome page as they are already logged
in